:sparkles: feat: start login screen
